+++ src/main/java/org/corpus_tools/annis/gui/ServiceStarterDesktop.java
@@ -32,6 +32,7 @@ import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
 import javax.swing.plaf.nimbus.NimbusLookAndFeel;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.SystemUtils;
 import org.corpus_tools.annis.gui.security.DesktopAuthentication;
 import org.corpus_tools.annis.gui.security.SecurityConfiguration;
 import org.slf4j.Logger;
@@ -202,8 +203,7 @@ public class ServiceStarterDesktop extends ServiceStarter { // NO_UCD (unused co
     claims.put("sub", USER_NAME);
     DefaultOAuth2User user = new DefaultOAuth2User(grantedAuthorities, claims, "sub");
 
-    this.desktopUserCredentials =
-        Optional.of(new DesktopAuthentication(user, signedToken));
+    this.desktopUserCredentials = Optional.of(new DesktopAuthentication(user, signedToken));
 
     // Open the application in the browser
     String webURL = "http://localhost:" + serverPort;
@@ -214,6 +214,24 @@ public class ServiceStarterDesktop extends ServiceStarter { // NO_UCD (unused co
       log.warn(
           "ANNIS is running in desktop mode, but no desktop has been detected. You can open {} manually.",
           webURL);
+    } else if (!("amd64".equals(SystemUtils.OS_ARCH) || "x86_64".equals(SystemUtils.OS_ARCH))) {
+      if (SystemUtils.IS_OS_MAC_OSX && "aarch64".equals(SystemUtils.OS_ARCH)) {
+        JOptionPane.showMessageDialog(null,
+            "ANNIS can only be run on Apple M1 systems using the Rosetta emulator "
+                + "(https://support.apple.com/en-us/HT211861). "
+                + "You can try to install and use the Adoptium OpenJDK Java 11 from "
+                + "https://adoptium.net to run ANNIS.",
+            "Cannot run ANNIS on incompatible operating system", JOptionPane.ERROR_MESSAGE);
+      } else {
+        JOptionPane.showMessageDialog(null,
+            "ANNIS can only be run on 64 bit operating systems (\"amd64\" or \"x86_64\") "
+                + "and with a 64 bit version of Java, "
+                + "but this computer is reported as architecture " + SystemUtils.OS_ARCH + "!\n\n"
+                + "A common cause is that a 32 bit version of Java is installed. "
+                + "Make sure to install a 64 bit Java 11 e.g. from https://adoptium.net",
+            "Cannot run ANNIS on incompatible operating system", JOptionPane.ERROR_MESSAGE);
+      }
+      System.exit(64);
     } else {
       showApplicationWindow();
       openBrowser(webURL);